Optimized word info download, two engines, fix bugs
Due to using multi-process to download, lock is duplicated to
two, so prone to meet two processes access the same web host
at the same time, this will be treated as DoS attack, server
will refuse to serve us. So, now use two engines for two processes
problem seems solved. And for some words like pencil-case, it is
not composed all by letters, so delete the isalpha() check